Created like/dislike section and date update functionality

+++ app/public/deleteMovie.php
@@ -3,6 +3,12 @@ session_start();
 
 require_once("config.php");
 
+// Check if the user is logged in, if not then redirect him to login page
+if (!isset($_SESSION["loggedin"]) || $_SESSION["loggedin"] !== true) {
+    header("location: index.php");
+    exit;
+}
+
 $userId = $_SESSION['id'];
 $movieId = $_GET['id'];
 $sql = 'DELETE FROM movies WHERE id = ' . $movieId . ' AND user_id = ' . $userId;
